trulens-backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
python-dotenv==1.0.1
httpx==0.27.2
pydantic[email]==2.9.2
python-multipart==0.0.12
